feat(core): emit addedUserIds on invitation accept and experience-flow jit
