+++ packages/core/src/routes/experience/classes/libraries/provision-library.ts
@@ -20,6 +20,7 @@ import { generateStandardId } from '@logto/shared';
 import { condArray, conditional, conditionalArray, trySafe } from '@silverhand/essentials';
 
 import { EnvSet } from '#src/env-set/index.js';
+import { truncateMembershipDelta } from '#src/libraries/hook/utils.js';
 import type TenantContext from '#src/tenants/TenantContext.js';
 import { buildAppInsightsTelemetry } from '#src/utils/request.js';
 import { getTenantId } from '#src/utils/tenant.js';
@@ -154,6 +155,8 @@ export class ProvisionLibrary {
     for (const { organizationId } of provisionedOrganizations) {
       this.ctx.appendDataHookContext('Organization.Membership.Updated', {
         organizationId,
+        ...(EnvSet.values.isDevFeaturesEnabled &&
+          truncateMembershipDelta({ addedUserIds: [payload.userId] })),
       });
     }
