+++ packages/core/src/routes/organization-invitation/index.ts
@@ -6,7 +6,9 @@ import {
 } from '@logto/schemas';
 import { z } from 'zod';
 
+import { EnvSet } from '#src/env-set/index.js';
 import RequestError from '#src/errors/RequestError/index.js';
+import { buildManagementApiContext, truncateMembershipDelta } from '#src/libraries/hook/utils.js';
 import koaGuard from '#src/middleware/koa-guard.js';
 import SchemaRouter from '#src/utils/SchemaRouter.js';
 import assertThat from '#src/utils/assert-that.js';
@@ -158,14 +160,42 @@ export default function organizationInvitationRoutes<T extends ManagementApiRout
         })
       );
 
+      // Snapshot pre-existing membership only when dev features are on. In production
+      // (dev-OFF) the webhook payload is the legacy `{ organizationId }`-only shape,
+      // so the snapshot would be unused DB work. The snapshot must run before
+      // `updateStatus` — the underlying insert uses ON CONFLICT DO NOTHING, so after
+      // it runs we cannot distinguish a real addition from a re-accept of an existing
+      // member. Returns the accurate `addedUserIds` for the webhook payload:
+      // `[acceptedUserId]` on a real addition, `[]` on a re-accept.
+      const addedUserIds = await (async (): Promise<string[]> => {
+        if (!EnvSet.values.isDevFeaturesEnabled) {
+          return [];
+        }
+        const invitation = await queries.organizations.invitations.findById(id);
+        const existingUserIds = await queries.organizations.relations.users.getExistingUserIds(
+          invitation.organizationId,
+          [acceptedUserId]
+        );
+        return existingUserIds.length > 0 ? [] : [acceptedUserId];
+      })();
+
       const result = await organizationInvitations.updateStatus(id, status, acceptedUserId);
 
+      ctx.body = result;
+      // Set explicitly so `buildManagementApiContext(ctx)` captures 200 rather than the
+      // koa-default 404 in the webhook payload.
+      ctx.status = 200;
+
+      // Always emit, matching every other Organization.Membership.Updated trigger in
+      // this project: the dev-features gate lives inside the spread, and
+      // `truncateMembershipDelta` omits empty arrays so a re-accept (no real change)
+      // produces the legacy `{ organizationId }`-only shape, identical to dev-OFF.
       const { organizationId } = result;
       ctx.appendDataHookContext('Organization.Membership.Updated', {
+        ...buildManagementApiContext(ctx),
         organizationId,
+        ...(EnvSet.values.isDevFeaturesEnabled && truncateMembershipDelta({ addedUserIds })),
       });
-
-      ctx.body = result;
 
       return next();
     }
